Add Stamp Switcher
The primary form is set to the placeholder stamp and we have to switch
it to use the stamp that was saved with the submission.

Also commented out the code that attempted to make this page work as a
seamless child (since we are not rendering it in an iframe).