Finish "elegant hack" for stamp
We were able to dynamically switch the stamp to use on the formio.js
side, but due to a server code base issue, the chosen stamp and its
data were not saving to the formio platform.

Travis is actually planning to fix it on the server side within a week
or so, so we’ll want to revisit this to see if it is necessary anymore.
 If not necessary, let’s get rid of the hack on all sides BEFORE
rolling to production.

On the formio.js side, we copied the dynamically chosen stamp’s data
into a hidden field (stampData) on the primary form during the
beforeSubmit hook.

On this side (Formio-Viewer), upon the submissionLoad event, we get a
handle on the nested form component (stamp) and reset its ‘form’
property.

From portal, we added a calculated value to the nested form component
set to run only in the viewer:
if ( typeof getValueFrom == ‘undefined’ ) {
     value = data.stampData;
}